+++ notanorm/__init__.py
@@ -16,7 +16,7 @@ except Exception:  # pragma: no cover
     pass
 
 from .sqlite import SqliteDb
-from .base import DbBase, Op, ReconnectionArgs
+from .base import DbBase, Op, Where, ReconnectionArgs
 from .model import (
     DbType,
     DbCol,
@@ -50,6 +50,7 @@ __all__ = [
     "DbIndexField",
     "errors",
     "Op",
+    "Where",
     "open_db",
     "model_from_ddl",
     "ReconnectionArgs",
