+++ tests/test_notanorm.py
@@ -3,22 +3,19 @@
 
 import copy
 import logging
-import multiprocessing
 import sqlite3
 import threading
 import time
-from multiprocessing.pool import ThreadPool, Pool as ProcessPool
+from multiprocessing.pool import ThreadPool
 from typing import Generator
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock
 
 import pytest
 
 import notanorm.errors
 import notanorm.errors as err
-from notanorm import SqliteDb, DbRow, DbBase, DbType, ReconnectionArgs
+from notanorm import SqliteDb, DbRow, DbBase, DbType, Where, Op
 from notanorm.connparse import open_db, parse_db_uri
-
-from tests.conftest import cleanup_mysql_db
 
 log = logging.getLogger(__name__)
 
@@ -388,44 +385,6 @@ def test_db_upsert_non_null(db):
     assert db.select_one("foo").bop == "keep"
 
 
-def test_conn_retry(db):
-    db.query("create table foo (x integer)")
-    db._conn_p.close()  # pylint: disable=no-member
-    db.max_reconnect_attempts = 1
-    with pytest.raises(Exception):
-        db.query("create table foo (x integer)")
-    db.max_reconnect_attempts = 2
-    db.query("create table bar (x integer)")
-    db.max_reconnect_attempts = 99
-
-
-def test_reconnect_cb(tmp_path):
-    def cb():
-        nonlocal cb_called
-        cb_called = True
-        raise Exception("ensure that the cb exception is caught")
-
-    cb_called = False
-    recon_args = ReconnectionArgs(max_reconnect_attempts=2, failure_callback=cb)
-    db = SqliteDb(str(tmp_path / "db"), reconnection_args=recon_args, timeout=0)
-    try:
-        db.query("create table foo (x integer)")
-        assert not cb_called
-        with patch.object(db, "_executeone", side_effect=err.DbConnectionError):
-            with pytest.raises(err.DbConnectionError):
-                db.query("create table bar (x integer)")
-    finally:
-        db.close()
-
-
-def test_conn_reopen(db):
-    db.query("create table foo (x integer)")
-    db.close()
-    assert db.closed
-    with pytest.raises(Exception):
-        db.query("create table foo (x integer)")
-
-
 def test_multi_close(db):
     db.close()
     db.close()
@@ -454,102 +413,6 @@ def test_safedb_inmemorydb(db):
     assert db.query("select bar from foo")[0].bar == 100
 
 
-def get_db(db_name, db_conn):
-    if db_name == "sqlite":
-        return SqliteDb(*db_conn[0], **db_conn[1])
-    else:
-        from notanorm import MySqlDb
-
-        return MySqlDb(*db_conn[0], **db_conn[1])
-
-
-def cleanup_db(db):
-    if db.__class__.__name__ == "MySqlDb":
-        cleanup_mysql_db(db)
-
-
-def _test_upsert_i(db_name, i, db_conn, mod):
-    db = get_db(db_name, db_conn)
-
-    with db.transaction() as db:
-        db.upsert("foo", bar=i % mod, baz=i)
-        row = db.select_one("foo", bar=i % mod)
-        db.update("foo", bar=i % mod, cnt=row.cnt + 1)
-
-
-# todo: maybe using the native "mysql connector" would enable fixing this
-#       but really, why would mysql allow blocking transactions like sqlite?
-@pytest.mark.db("sqlite")
-def test_upsert_multiprocess(db_name, db_notmem, tmp_path):
-    db = db_notmem
-    create_and_fill_test_db(db, 0)
-
-    num = 22
-    ts = []
-    mod = 5
-
-    for i in range(0, num):
-        currt = multiprocessing.Process(
-            target=_test_upsert_i,
-            args=(db_name, i, db.connection_args, mod),
-            daemon=True,
-        )
-        ts.append(currt)
-
-    for t in ts:
-        t.start()
-
-    for t in ts:
-        t.join()
-
-    log.debug(db.select("foo"))
-    assert len(db.select("foo")) == mod
-    for i in range(mod):
-        ent = db.select_one("foo", bar=i)
-        assert ent.cnt == int(num / mod) + (i < num % mod)
-
-
-# for some reqson mysql seems connect in a way that causes multiple object to have the same underlying connection
-# todo: maybe using the native "mysql connector" would enable fixing this
-@pytest.mark.db("sqlite")
-def test_upsert_threaded_multidb(db_notmem, db_name):
-    print("sqlite3 version", sqlite3.sqlite_version)
-
-    db = db_notmem
-    db.query(
-        "create table foo (bar integer primary key, baz integer, cnt integer default 0)"
-    )
-
-    num = 22
-    ts = []
-    mod = 5
-
-    def _upsert_i(up_i, up_db_name, db_conn, up_mod):
-        with get_db(up_db_name, db_conn) as _db:
-            _db.upsert("foo", bar=up_i % up_mod, baz=up_i)
-            with _db.transaction():
-                row = _db.select_one("foo", bar=up_i % up_mod)
-                _db.update("foo", bar=up_i % up_mod, cnt=row.cnt + 1)
-
-    for i in range(0, num):
-        currt = threading.Thread(
-            target=_upsert_i, args=(i, db_name, db.connection_args, mod), daemon=True
-        )
-        ts.append(currt)
-
-    for t in ts:
-        t.start()
-
-    for t in ts:
-        t.join()
-
-    log.debug(db.select("foo"))
-    assert len(db.select("foo")) == mod
-    for i in range(mod):
-        ent = db.select_one("foo", bar=i)
-        assert ent.cnt == int(num / mod) + (i < num % mod)
-
-
 def test_transactions_any_exc(db):
     class TestExc(Exception):
         pass
@@ -576,38 +439,6 @@ def test_transactions_deadlock(db):
     for i in range(50):
         db.insert("foo", bar=i)
     thread.join()
-
-
-def test_upsert_thready_one(db_notmem):
-    db = db_notmem
-
-    create_and_fill_test_db(db, 0)
-
-    failed = False
-    num = 100
-    mod = 7
-
-    def upsert_i(db, i):
-        try:
-            db.upsert("foo", bar=str(i % mod), baz=i)
-        except Exception as e:
-            nonlocal failed
-            failed = True
-            log.error("failed to upsert: %s", repr(e))
-
-    ts = []
-    for i in range(0, num):
-        currt = threading.Thread(target=upsert_i, args=(db, i), daemon=True)
-        ts.append(currt)
-
-    for t in ts:
-        t.start()
-
-    for t in ts:
-        t.join()
-
-    assert not failed
-    assert len(db.select("foo")) == mod
 
 
 def test_select_gen_not_lock(db: DbBase):
@@ -653,23 +484,6 @@ def test_select_gen_not_lock(db: DbBase):
     assert (end - start) < 1
 
 
-# for some reqson mysql seems connect in a way that causes multiple object to have the same underlying connection
-# todo: maybe using the native "connector" would enable fixing this
-@pytest.mark.db("sqlite")
-def test_transaction_fail_on_begin(db_notmem: "DbBase", db_name):
-    print("sqlite3 version", sqlite3.sqlite_version)
-
-    db1 = db_notmem
-    db2 = get_db(db_name, db1.connection_args)
-
-    db1.max_reconnect_attempts = 1
-
-    with db2.transaction():
-        with pytest.raises(sqlite3.OperationalError, match=r".*database.*is locked"):
-            with db1.transaction():
-                pass
-
-
 def test_readonly_fail(db, db_name: str):
     db.query("create table foo (bar text)")
     db.insert("foo", bar="y1")
@@ -683,49 +497,6 @@ def test_readonly_fail(db, db_name: str):
 
     with pytest.raises(err.DbReadOnlyError):
         db.insert("foo", bar="y2")
-
-
-@pytest.mark.db("sqlite")
-def test_collation(db):
-    def collate(v1, v2):
-        return 1 if v1 > v2 else -1 if v1 < v2 else 0
-
-    db._conn().create_collation("COMP", collate)
-    db.query("create table foo (bar text collate COMP)")
-    # collation + multi-thread mode = possible deadlock
-    db.use_collation_locks = True
-
-    with db.transaction():
-        for i in range(5000):
-            db.insert("foo", bar=str(i))
-
-    evt = threading.Event()
-
-    def select_gen():
-        evt.wait()
-        log.warning("gen-start")
-        for row in db.select_gen("foo", order_by="bar"):
-            log.info("gen: %s", row.bar)
-        log.warning("gen-end")
-
-    def select_one():
-        evt.wait()
-        log.warning("one-start")
-        for i in range(5000):
-            row = db.select_one("foo", bar=str(i))
-            log.info("one: %s", row.bar)
-        log.warning("one-end")
-
-    select_gen_thread = threading.Thread(target=select_gen, daemon=True)
-    select_one_thread = threading.Thread(target=select_one, daemon=True)
-
-    select_gen_thread.start()
-    select_one_thread.start()
-
-    evt.set()
-
-    select_gen_thread.join()
-    select_one_thread.join()
 
 
 def test_missing_column(db):
@@ -950,19 +721,6 @@ def test_sqlite_guard_thread(db_notmem):
         break
 
     assert cool
-
-
-def upserty(uri, i):
-    db = open_db(uri)
-    try:
-        db.generator_guard = True
-        for row in db.select_gen("foo"):
-            db.upsert("foo", bar=row.bar, baz=row.baz + 1)
-        # this is ok: we passed
-        return i
-    except err.UnsafeGeneratorError:
-        # this is ok: we created a consistent error
-        return -1
 
 
 def test_subq(db):
@@ -1210,6 +968,23 @@ def test_where_or(db):
     assert len(db.select("foo")) == 3
 
 
+def test_where_complex(db):
+    create_and_fill_test_db(db, 5)
+    assert (
+        len(db.select("foo", _where=Where([{"bar": Op(">", 1)}, {"bar": Op("<", 5)}])))
+        == 3
+    )
+    assert (
+        len(
+            db.select(
+                "foo",
+                _where=[{"bar": 1}, Where([{"bar": Op(">", 1)}, {"bar": Op("<", 5)}])],
+            )
+        )
+        == 4
+    )
+
+
 def test_del_raises(db):
     create_and_fill_test_db(db, 5)
     db.delete("foo", bar=2)
@@ -1218,31 +993,6 @@ def test_del_raises(db):
         db.delete("foo")
     with pytest.raises(ValueError):
         db.delete("foo", {"bar": 3}, baz=0)
-
-
-def test_generator_proc(db_notmem):
-    db = db_notmem
-
-    uri = db.uri
-    log.debug("using uri" + uri)
-
-    create_and_fill_test_db(db, 20)
-    db.close()
-
-    proc_num = 4
-
-    pool = ProcessPool(processes=proc_num)
-
-    import functools
-
-    func = functools.partial(upserty, uri)
-
-    expected = list(range(proc_num * 2))
-
-    if db.uri_name == "sqlite":
-        expected = [-1] * proc_num * 2
-
-    assert pool.map(func, range(proc_num * 2)) == expected
 
 
 def test_db_direct_clone(db_notmem):
